pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "xybernetex"
version = "0.1.0"
description = "Python SDK for the Xybernetex autonomous agent platform"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "requests>=2.31",
    "httpx>=0.27",
]

[project.optional-dependencies]
docx = ["python-docx>=1.1"]
pdf  = ["reportlab>=4.0", "markdown>=3.5"]
xlsx = ["openpyxl>=3.1"]
pptx = ["python-pptx>=0.6"]
all  = [
    "python-docx>=1.1",
    "reportlab>=4.0",
    "markdown>=3.5",
    "openpyxl>=3.1",
    "python-pptx>=0.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["xybernetex*"]
